+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 only, not configurable via UI anymore and existing values not taken into account.
 See ldap.ini.example if needed or doc for env variables.
 
+* [booking] avoid any resource selection in resources input
 * [booking] Fix reservation form unexpected submissions
 * [booking] fix bk_scheduling fetching from booking view
 * [self_registration] add structures names to listed spaces